move buffer from iteration-utils to its own class
